backend/requirements.txt
asgiref==3.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1.4
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-filter==22.1
django-link-shortener==0.5
django-rest-framework==0.1.0
django-taggit==3.1.0
django-urlshortener==2.0.0
djangoadmin==2.2.8
djangoarticle==2.2.4
djangocomment==2.2.2
djangocontact==2.2.3
djangopost==2.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangotags==2.2.2
djangotools==2.2.2
drf-yasg==1.21.4
flake8==6.0.0
freezegun==1.2.2
idna==3.4
inflection==0.5.1
isort==5.11.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
mysqlclient==2.1.1
oauthlib==3.2.2
packaging==22.0
Pillow==9.3.0
pip==22.0.3
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pyshorteners==1.0.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
setuptools==60.6.0
six==1.16.0
sqlparse==0.4.3
test-plus==0.2.2
uritemplate==4.1.1
urllib3==1.26.13
wheel==0.37.1
